feat: embed remote desktop workspace in tool dialog
